+++ packages/expo-modules-core/ios/AppDelegates/ExpoAppDelegate.swift
@@ -115,21 +115,28 @@ open class ExpoAppDelegate: UIResponder, UIApplicationDelegate {
     let selector = #selector(application(_:didReceiveRemoteNotification:fetchCompletionHandler:))
     let subs = subscribers.filter { $0.responds(to: selector) }
     var subscribersLeft = subs.count
-    var fetchResult: UIBackgroundFetchResult = .noData
     let dispatchQueue = DispatchQueue(label: "expo.application.remoteNotification", qos: .userInteractive)
+    var failedCount = 0
+    var newDataCount = 0
 
     let handler = { (result: UIBackgroundFetchResult) in
       dispatchQueue.sync {
         if result == .failed {
-          fetchResult = .failed
-        } else if fetchResult != .failed && result == .newData {
-          fetchResult = .newData
+          failedCount += 1
+        } else if result == .newData {
+          newDataCount += 1
         }
 
         subscribersLeft -= 1
 
         if subscribersLeft == 0 {
-          completionHandler(fetchResult)
+          if newDataCount > 0 {
+            completionHandler(.newData)
+          } else if failedCount > 0 {
+            completionHandler(.failed)
+          } else {
+            completionHandler(.noData)
+          }
         }
       }
     }
@@ -216,21 +223,28 @@ open class ExpoAppDelegate: UIResponder, UIApplicationDelegate {
     let selector = #selector(application(_:performFetchWithCompletionHandler:))
     let subs = subscribers.filter { $0.responds(to: selector) }
     var subscribersLeft = subs.count
-    var fetchResult: UIBackgroundFetchResult = .noData
     let dispatchQueue = DispatchQueue(label: "expo.application.performFetch", qos: .userInteractive)
+    var failedCount = 0
+    var newDataCount = 0
 
     let handler = { (result: UIBackgroundFetchResult) in
       dispatchQueue.sync {
         if result == .failed {
-          fetchResult = .failed
-        } else if fetchResult != .failed && result == .newData {
-          fetchResult = .newData
+          failedCount += 1
+        } else if result == .newData {
+          newDataCount += 1
         }
 
         subscribersLeft -= 1
 
         if subscribersLeft == 0 {
-          completionHandler(fetchResult)
+          if newDataCount > 0 {
+             completionHandler(.newData)
+           } else if failedCount > 0 {
+             completionHandler(.failed)
+           } else {
+             completionHandler(.noData)
+           }
         }
       }
     }